refactor: simplify task dependency registration in bun.ts and node.ts
- Updated the task registration process by replacing the previous implementation with streamlined server functions: `registerMcpTaskDepsServer` and `registerBrowserDepsServer`.
- Adjusted the TypeScript configuration to include all source files under the `src` directory for better coverage.
