+++ packages/tasks/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src/common.ts", "src/*/**/*"],
+  "include": ["src/**/*"],
   "files": ["./src/browser.ts", "./src/node.ts", "./src/bun.ts"],
   "exclude": ["dist", "node_modules"],
   "compilerOptions": {
